Pipeline_HDFS_SPARK/scripts/requirements.txt
# Big Data Pipeline - Python Bağımlılıkları

# Kafka
kafka-python>=2.0.2

# PySpark
pyspark>=3.5.0

# MongoDB
pymongo>=4.6.0

# Kaggle API
kaggle>=1.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
